Move rotator state definition to include/hamlib/rot_state.h
